fix: RateRepository 구현객체 부분 분리
